+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unrelease]
+### Added
+- `setIsolationLevel` for transactions with `DAO` and `Database` in Mysql/MariaDb.
 ### Fixed
 - `Injected` attribute parsing in `DiControllerParser` (pb with name) - no open issue -